% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMethod.R
\name{cmColDef}
\alias{cmColDef}
\title{cmColDef}
\usage{
cmColDef(elementId = "cm-tc-tab", colNames = NULL)
}
\arguments{
\item{elementId}{An element id for the table using this column definitions (needed for the drop down selection)}

\item{colNames}{Optional a vector of column names to restrict to}
}
\value{
A column definition list
}
\description{
Extract cohorts from json
}
\details{
Returns a names list with the cohorts
}
\seealso{
Other ColDefs:
\code{\link[=cmOutcomeColDef]{cmOutcomeColDef()}},
\code{\link[=defaultColumns]{defaultColumns()}},
\code{\link[=getCIcolumns]{getCIcolumns()}},
\code{\link[=getCdCols]{getCdCols()}},
\code{\link[=getPlpColDefs]{getPlpColDefs()}},
\code{\link[=getSccsColDefs]{getSccsColDefs()}}
}
\concept{ColDefs}
