% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMethod.R
\name{cmOutcomeColDef}
\alias{cmOutcomeColDef}
\title{cmOutcomeColDef}
\usage{
cmOutcomeColDef(colNames)
}
\arguments{
\item{colNames}{Optional a vector of column names to restrict to}
}
\value{
A column definition list
}
\description{
List with column names for the cohort method outcome table
}
\details{
Returns a names list with the cohorts names
}
\seealso{
Other ColDefs:
\code{\link[=cmColDef]{cmColDef()}},
\code{\link[=defaultColumns]{defaultColumns()}},
\code{\link[=getCIcolumns]{getCIcolumns()}},
\code{\link[=getCdCols]{getCdCols()}},
\code{\link[=getPlpColDefs]{getPlpColDefs()}},
\code{\link[=getSccsColDefs]{getSccsColDefs()}}
}
\concept{ColDefs}
