% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortIncidence.R
\name{createStratSentance}
\alias{createStratSentance}
\title{createStratSentance}
\usage{
createStratSentance(CohortIncidenceModuleSettings)
}
\arguments{
\item{CohortIncidenceModuleSettings}{The cohort incidence module specification}
}
\value{
An string with the stratification information
}
\description{
Create a sentence that explains the stratification defined in the analysis.
}
\details{
Returns a string
}
