% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{formatCovariateSettings}
\alias{formatCovariateSettings}
\title{formatCovariateSettings}
\usage{
formatCovariateSettings(covariateSettings)
}
\arguments{
\item{covariateSettings}{The covariate settings}
}
\value{
a data.frame with the covariate settings
}
\description{
changes the covariateSettings list to a nice table format
}
\details{
Returns a data.frame with the covariate settings
}
\seealso{
Other Helpers:
\code{\link[=functionDefaults]{functionDefaults()}},
\code{\link[=getAllHelpDetails]{getAllHelpDetails()}},
\code{\link[=getAllHelpText]{getAllHelpText()}},
\code{\link[=getCohortDefinitionsFromJson]{getCohortDefinitionsFromJson()}},
\code{\link[=getConcepts]{getConcepts()}},
\code{\link[=getDemoLoc]{getDemoLoc()}},
\code{\link[=getFunctionFromArgName]{getFunctionFromArgName()}},
\code{\link[=getHelpText]{getHelpText()}},
\code{\link[=getNegativeControlsFromJson]{getNegativeControlsFromJson()}},
\code{\link[=getSettingsTable]{getSettingsTable()}},
\code{\link[=reportTableFormat]{reportTableFormat()}},
\code{\link[=tagPrint]{tagPrint()}}
}
\concept{Helpers}
