% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDiagnostics.R
\name{getCdCols}
\alias{getCdCols}
\title{getCdCols}
\usage{
getCdCols()
}
\value{
A list of colDefs for cohort diagnostic
}
\description{
A list of colDefs for the columns used in cohort diagnostic protocol module
}
\details{
Returns a list of colDefs
}
