% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortIncidence.R
\name{getCiTargetsOutcomes}
\alias{getCiTargetsOutcomes}
\title{getCiTargetsOutcomes}
\usage{
getCiTargetsOutcomes(CohortIncidenceModuleSettings, cohortDefinitionDf)
}
\arguments{
\item{CohortIncidenceModuleSettings}{The cohort incidence module specification}

\item{cohortDefinitionDf}{The data.frame with the cohort definition details}
}
\value{
A list with the tables to present in the protocol
}
\description{
Extracts a list of target tables, outcome tables, TARs and vector of unique targetIds and outcomeIds in analysis
}
\details{
Returns a list
}
