% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{getCohortDefinitionsFromJson}
\alias{getCohortDefinitionsFromJson}
\title{getCohortDefinitionsFromJson}
\usage{
getCohortDefinitionsFromJson(json)
}
\arguments{
\item{json}{The json analysis specification}
}
\value{
An named R list with the elements subsetUnique (list of subset operators), cohortDefinitions (list of cohortDefinitions) and cohortDefinitionDf (data.frame of cohort definitions)
}
\description{
Extract cohorts from json
}
\details{
Returns a names list with the cohorts
}
\seealso{
Other Helpers:
\code{\link[=formatCovariateSettings]{formatCovariateSettings()}},
\code{\link[=functionDefaults]{functionDefaults()}},
\code{\link[=getAllHelpDetails]{getAllHelpDetails()}},
\code{\link[=getAllHelpText]{getAllHelpText()}},
\code{\link[=getConcepts]{getConcepts()}},
\code{\link[=getDemoLoc]{getDemoLoc()}},
\code{\link[=getFunctionFromArgName]{getFunctionFromArgName()}},
\code{\link[=getHelpText]{getHelpText()}},
\code{\link[=getNegativeControlsFromJson]{getNegativeControlsFromJson()}},
\code{\link[=getSettingsTable]{getSettingsTable()}},
\code{\link[=reportTableFormat]{reportTableFormat()}},
\code{\link[=tagPrint]{tagPrint()}}
}
\concept{Helpers}
