% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDiagnostics.R
\name{getCohortDiagnosticTables}
\alias{getCohortDiagnosticTables}
\title{getCohortDiagnosticTables}
\usage{
getCohortDiagnosticTables(CohortDiagnosticsSettings, cohortDefinitionDf)
}
\arguments{
\item{CohortDiagnosticsSettings}{The cohort diagnostic module specification}

\item{cohortDefinitionDf}{The data.frame with the cohort definition details}
}
\value{
A list with the tables to display
}
\description{
Extract target, setting and feature table from CohortDiagnosticsSettings
}
\details{
Returns a list of tables
}
