% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{getHelpText}
\alias{getHelpText}
\title{getHelpText}
\usage{
getHelpText(
  package = "CohortMethod",
  functionName = "createPs",
  input = "maxCohortSizeForFitting"
)
}
\arguments{
\item{package}{Name of installed R package of interest}

\item{functionName}{Name of exported function inside R package of interest}

\item{input}{Name of input into exported function}
}
\value{
Details about the input
}
\description{
For an installed R package, user can specify a function in the package
and in input for the function to get details about the input.
}
\details{
Returns a string with details about the input
}
\seealso{
Other Helpers:
\code{\link[=formatCovariateSettings]{formatCovariateSettings()}},
\code{\link[=functionDefaults]{functionDefaults()}},
\code{\link[=getAllHelpDetails]{getAllHelpDetails()}},
\code{\link[=getAllHelpText]{getAllHelpText()}},
\code{\link[=getCohortDefinitionsFromJson]{getCohortDefinitionsFromJson()}},
\code{\link[=getConcepts]{getConcepts()}},
\code{\link[=getDemoLoc]{getDemoLoc()}},
\code{\link[=getFunctionFromArgName]{getFunctionFromArgName()}},
\code{\link[=getNegativeControlsFromJson]{getNegativeControlsFromJson()}},
\code{\link[=getSettingsTable]{getSettingsTable()}},
\code{\link[=reportTableFormat]{reportTableFormat()}},
\code{\link[=tagPrint]{tagPrint()}}
}
\concept{Helpers}
