% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatientLevelPrediction.R
\name{getPlpColDefs}
\alias{getPlpColDefs}
\title{getPlpColDefs}
\usage{
getPlpColDefs()
}
\value{
A column definition list
}
\description{
create colDefs for prediction table
}
\details{
Returns a names list with the cohorts
}
