% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfControlCaseSeries.R
\name{getSccsColDefs}
\alias{getSccsColDefs}
\title{getSccsColDefs}
\usage{
getSccsColDefs()
}
\value{
A column definition list
}
\description{
Create col defs for SCCS
}
\details{
Returns a named list of colDefs
}
\seealso{
Other ColDefs:
\code{\link[=cmColDef]{cmColDef()}},
\code{\link[=cmOutcomeColDef]{cmOutcomeColDef()}},
\code{\link[=defaultColumns]{defaultColumns()}},
\code{\link[=getCIcolumns]{getCIcolumns()}},
\code{\link[=getCdCols]{getCdCols()}},
\code{\link[=getPlpColDefs]{getPlpColDefs()}}
}
\concept{ColDefs}
