% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfControlCaseSeries.R
\name{getSccsSettings}
\alias{getSccsSettings}
\title{getSccsSettings}
\usage{
getSccsSettings(
  SelfControlledCaseSeriesModuleSettings,
  cohortDefinitionDf,
  negativeControls
)
}
\arguments{
\item{SelfControlledCaseSeriesModuleSettings}{The self controlled case series module specification}

\item{cohortDefinitionDf}{The data.frame with the cohort definition details}

\item{negativeControls}{The shared negative controls from the json spec}
}
\value{
A list with the tables to display
}
\description{
Extract exposure/outcomes of interest, shared and analysis specific negative controls, diagnostic settings and analysis settings
}
\details{
Returns a list of tables and settings
}
