% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{reportTableFormat}
\alias{reportTableFormat}
\title{reportTableFormat}
\usage{
reportTableFormat(
  table,
  groupBy = NULL,
  columns = NULL,
  elementId = NULL,
  caption,
  groupByButton = FALSE
)
}
\arguments{
\item{table}{data.frame or tibble with the data to present}

\item{groupBy}{column to group by (optional)}

\item{columns}{The column details (create default using defaultColumns())}

\item{elementId}{Element ID for the widget.}

\item{caption}{A table caption}
}
\value{
Details about all inputs into the functionName within R package of interest
}
\description{
create a grouped reactable::reactable
}
\details{
Returns a reactable::reactable
}
