% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{tagPrint}
\alias{tagPrint}
\title{display a reactable when there are multiple outputs to print}
\usage{
tagPrint(x)
}
\arguments{
\item{x}{The object, such as a reactable, to print}
}
\value{
Nothing just prints the object in quarto
}
\description{
display a reactable when there are multiple outputs to print
}
\details{
Wraps the input around print and shiny:tagList
}
\seealso{
Other Helpers:
\code{\link[=formatCovariateSettings]{formatCovariateSettings()}},
\code{\link[=functionDefaults]{functionDefaults()}},
\code{\link[=getAllHelpDetails]{getAllHelpDetails()}},
\code{\link[=getAllHelpText]{getAllHelpText()}},
\code{\link[=getCohortDefinitionsFromJson]{getCohortDefinitionsFromJson()}},
\code{\link[=getConcepts]{getConcepts()}},
\code{\link[=getDemoLoc]{getDemoLoc()}},
\code{\link[=getFunctionFromArgName]{getFunctionFromArgName()}},
\code{\link[=getHelpText]{getHelpText()}},
\code{\link[=getNegativeControlsFromJson]{getNegativeControlsFromJson()}},
\code{\link[=getSettingsTable]{getSettingsTable()}},
\code{\link[=reportTableFormat]{reportTableFormat()}}
}
\concept{Helpers}
