% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{defaultColumns}
\alias{defaultColumns}
\title{defaultColumns}
\usage{
defaultColumns(data)
}
\arguments{
\item{data}{the data.frame to use in reactable}
}
\value{
Returns a reactable colunn definition
}
\description{
create reactable colunn definition where
if the column bold is TRUE or NA the column
value had red font otherwise black
the bold column is then hidden and column level2
had no name shown
}
\details{
Returns a reactable colunn definition
}
