% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMethod.R
\name{extractCohortMethodSettings}
\alias{extractCohortMethodSettings}
\title{extractCohortMethodSettings}
\usage{
extractCohortMethodSettings(
  cohortMethodModuleSettings,
  negativeControls,
  cohortDefinitionDf
)
}
\arguments{
\item{cohortMethodModuleSettings}{The cohort method module specification}

\item{negativeControls}{NULL or a data.frame of the negative controls}

\item{cohortDefinitionDf}{The data.frame with the cohort definition details}
}
\value{
An named R list with ...
}
\description{
Extract cohorts from json
}
\details{
Returns a names list with the cohorts
}
