% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{formatCovariateSettings}
\alias{formatCovariateSettings}
\title{formatCovariateSettings}
\usage{
formatCovariateSettings(covariateSettings)
}
\arguments{
\item{covariateSettings}{The covariate settings}
}
\value{
a data.frame with the covariate settings
}
\description{
changes the covariateSettings list to a nice table format
}
\details{
Returns a data.frame with the covariate settings
}
