% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generateProtocol}
\alias{generateProtocol}
\title{generateProtocol}
\usage{
generateProtocol(
  jsonLocation,
  json,
  dataDiagnosticFolder = NULL,
  webAPI,
  authMethod = NULL,
  webApiUsername = NULL,
  webApiPassword = NULL,
  outputLocation = getwd(),
  outputName = paste0("protocol_", gsub(":", "_", gsub(" ", "_", as.character(date()))),
    ".html"),
  intermediateDir = tempdir(),
  downloadConcepts = TRUE,
  conceptsAsExcel = FALSE,
  conceptFolder = outputLocation,
  addCohortDefinitions = TRUE,
  exportCohortLocation = NULL,
  headerColor = "#336B91",
  headerLogoLocation = NULL,
  protocolSubheading = NULL
)
}
\arguments{
\item{jsonLocation}{The location of the json specification}

\item{json}{The json specification as an R list (this can be used instead of jsonLocation)}

\item{dataDiagnosticFolder}{The location of the database diagnostic results}

\item{webAPI}{the webAPI address}

\item{authMethod}{The webAPI authorization method (optional)}

\item{webApiUsername}{The webAPI authorization username (optional)}

\item{webApiPassword}{The webAPI authorization password (optional)}

\item{outputLocation}{The file location and name to save the protocol}

\item{outputName}{The name of the html protocol that is created}

\item{intermediateDir}{The work directory for quarto}

\item{downloadConcepts}{Whether to download the concepts from webAPI for the cohorts (can be slow)}

\item{conceptsAsExcel}{Whether to save the concepts into excel files rather than embed into html file if downloadConcepts is TRUE}

\item{conceptFolder}{The location to save the excel files if downloadConcepts is TRUE and conceptsAsExcel is TRUE}

\item{addCohortDefinitions}{Whether to add the cohorts to the protocol (can make document large)}

\item{exportCohortLocation}{if not NULL the location where the table tracter will be exported to csv.}

\item{headerColor}{The CSS color to use for the protocol header banner (optional)}

\item{headerLogoLocation}{The location of a logo image to add to the protocol header banner (optional)}

\item{protocolSubheading}{Optional subheading to show under the protocol title}
}
\value{
An named R list with the elements 'standard' and 'source'
}
\description{
Generates a protocol from a Strategus json specification
}
\details{
Specify the location of the json specification file, the data diagnostic folder (optional)
the ATLAS webAPI (to process cohorts and concepts) and where you want the
protocol saved.
}
