% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generateProtocol}
\alias{generateProtocol}
\title{generateProtocol}
\usage{
generateProtocol(
  jsonLocation = getDemoLoc(),
  dataDiagnosticFolder = NULL,
  webAPI,
  outputLocation = getwd(),
  outputName = paste0("protocol_", gsub(":", "_", gsub(" ", "_", as.character(date()))),
    ".html"),
  intermediateDir = tempdir()
)
}
\arguments{
\item{jsonLocation}{The location of the json specification}

\item{dataDiagnosticFolder}{The location of the database diagnostic results}

\item{webAPI}{the webAPI address}

\item{outputLocation}{The file location and name to save the protocol}

\item{outputName}{The name of the html protocol that is created}

\item{intermediateDir}{The work directory for quarto}
}
\value{
An named R list with the elements 'standard' and 'source'
}
\description{
Generates a protocol from a Strategus json specification
}
\details{
Specify the location of the json specification file, the data diagnostic folder (optional)
the ATLAS webAPI (to process cohorts and concepts) and where you want the
protocol saved.
}
