% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortIncidence.R
\name{getCIcolumns}
\alias{getCIcolumns}
\title{getCIcolumns}
\usage{
getCIcolumns()
}
\value{
A list of colDefs for the target and outcome tables describing the cohort incidence analysis
}
\description{
A reactable colDef list for the cohort incidence tables
}
\details{
Returns a list of colDefs
}
\seealso{
Other ColDefs:
\code{\link[=cmColDef]{cmColDef()}},
\code{\link[=cmOutcomeColDef]{cmOutcomeColDef()}},
\code{\link[=defaultColumns]{defaultColumns()}},
\code{\link[=getCdCols]{getCdCols()}},
\code{\link[=getPlpColDefs]{getPlpColDefs()}},
\code{\link[=getSccsColDefs]{getSccsColDefs()}}
}
\concept{ColDefs}
