% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{getCohortDefinitionsFromJson}
\alias{getCohortDefinitionsFromJson}
\title{getCohortDefinitionsFromJson}
\usage{
getCohortDefinitionsFromJson(json)
}
\arguments{
\item{json}{The json analysis specification}
}
\value{
An named R list with the elements subSetDefs, cohortIds, cohortNames and cohortDefinitions
}
\description{
Extract cohorts from json
}
\details{
Returns a names list with the cohorts
}
