% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{getConcepts}
\alias{getConcepts}
\title{getConcepts}
\usage{
getConcepts(
  expression = NULL,
  conceptIds = NULL,
  baseUrl = "https://api.ohdsi.org/WebAPI"
)
}
\arguments{
\item{expression}{A concept set expression to extract details about from ATLAS webapi}

\item{conceptIds}{A vector of conceptsIds to extract details about from ATLAS webapi}

\item{baseUrl}{The ATLAS baseUrl}
}
\value{
An named R list with the elements 'standard' and 'source'
}
\description{
Extract concept id details for a vector of concept ids
}
\details{
Returns a names list of length two with 'standard' and 'source' containing 
a data.frame with the concept ids details for the standard concepts and their
sourced concepts.
}
\seealso{
Other Helpers:
\code{\link[=formatCovariateSettings]{formatCovariateSettings()}},
\code{\link[=functionDefaults]{functionDefaults()}},
\code{\link[=getAllHelpDetails]{getAllHelpDetails()}},
\code{\link[=getAllHelpText]{getAllHelpText()}},
\code{\link[=getCohortDefinitionsFromJson]{getCohortDefinitionsFromJson()}},
\code{\link[=getDemoLoc]{getDemoLoc()}},
\code{\link[=getFunctionFromArgName]{getFunctionFromArgName()}},
\code{\link[=getHelpText]{getHelpText()}},
\code{\link[=getNegativeControlsFromJson]{getNegativeControlsFromJson()}},
\code{\link[=getSettingsTable]{getSettingsTable()}},
\code{\link[=reportTableFormat]{reportTableFormat()}},
\code{\link[=tagPrint]{tagPrint()}}
}
\concept{Helpers}
