% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{getConcepts}
\alias{getConcepts}
\title{getConcepts}
\usage{
getConcepts(conceptIds, baseUrl = "https://api.ohdsi.org/WebAPI")
}
\arguments{
\item{conceptIds}{A list of concept ids to extract details about from ATLAS webapi}

\item{baseUrl}{The ATLAS baseUrl}
}
\value{
An named R list with the elements 'standard' and 'source'
}
\description{
Extract concept id details for a vector of concept ids
}
\details{
Returns a names list of length two with 'standard' and 'source' containing 
a data.frame with the concept ids details for the standard concepts and their
sourced concepts.
}
