% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortIncidence.R
\name{getCountStatement}
\alias{getCountStatement}
\title{getCountStatement}
\usage{
getCountStatement(CohortIncidenceModuleSettings, cohortDefinitionDf)
}
\arguments{
\item{CohortIncidenceModuleSettings}{The cohort incidence module specification}

\item{cohortDefinitionDf}{The data.frame with the cohort definition details}
}
\value{
An string with the count information
}
\description{
Create a sentence that explains the number of targets, outcomes and settings per analysis.
}
\details{
Returns a string
}
