% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{getDemoLoc}
\alias{getDemoLoc}
\title{getDemoLoc}
\usage{
getDemoLoc()
}
\value{
A file path location to an example specification json
}
\description{
Get the file location of an example json in the package
}
\details{
This function returns the path to an example json specification file.
}
