% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{getNegativeControlsFromJson}
\alias{getNegativeControlsFromJson}
\title{getNegativeControlsFromJson}
\usage{
getNegativeControlsFromJson(json)
}
\arguments{
\item{json}{The json analysis specification}
}
\value{
A data.frame with the negative control details or NULL if no negative controls
}
\description{
Extract cohorts from json
}
\details{
Returns a names list with the cohorts
}
\seealso{
Other Helpers:
\code{\link[=formatCovariateSettings]{formatCovariateSettings()}},
\code{\link[=functionDefaults]{functionDefaults()}},
\code{\link[=getAllHelpDetails]{getAllHelpDetails()}},
\code{\link[=getAllHelpText]{getAllHelpText()}},
\code{\link[=getCohortDefinitionsFromJson]{getCohortDefinitionsFromJson()}},
\code{\link[=getConcepts]{getConcepts()}},
\code{\link[=getDemoLoc]{getDemoLoc()}},
\code{\link[=getFunctionFromArgName]{getFunctionFromArgName()}},
\code{\link[=getHelpText]{getHelpText()}},
\code{\link[=getSettingsTable]{getSettingsTable()}},
\code{\link[=reportTableFormat]{reportTableFormat()}},
\code{\link[=tagPrint]{tagPrint()}}
}
\concept{Helpers}
