% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatientLevelPrediction.R
\name{getPlpColDefs}
\alias{getPlpColDefs}
\title{getPlpColDefs}
\usage{
getPlpColDefs()
}
\value{
A column definition list
}
\description{
create colDefs for prediction table
}
\details{
Returns a names list with the cohorts
}
\seealso{
Other ColDefs:
\code{\link[=cmColDef]{cmColDef()}},
\code{\link[=cmOutcomeColDef]{cmOutcomeColDef()}},
\code{\link[=defaultColumns]{defaultColumns()}},
\code{\link[=getCIcolumns]{getCIcolumns()}},
\code{\link[=getCdCols]{getCdCols()}},
\code{\link[=getSccsColDefs]{getSccsColDefs()}}
}
\concept{ColDefs}
