% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatientLevelPrediction.R
\name{getPlpSettings}
\alias{getPlpSettings}
\title{getPlpSettings}
\usage{
getPlpSettings(PatientLevelPredictionModuleSettings, cohortDefinitionDf)
}
\arguments{
\item{PatientLevelPredictionModuleSettings}{The patient level prediction module specification}

\item{cohortDefinitionDf}{The data.frame with the cohort definition details}
}
\value{
An named R list with ...
}
\description{
Extract plp tables and settings from json
}
\details{
Returns a names list with the tables and settings
}
\seealso{
Other Extraction:
\code{\link[=createStratSentance]{createStratSentance()}},
\code{\link[=extractCohortMethodSettings]{extractCohortMethodSettings()}},
\code{\link[=getCiTargetsOutcomes]{getCiTargetsOutcomes()}},
\code{\link[=getCohortDiagnosticTables]{getCohortDiagnosticTables()}},
\code{\link[=getCountStatement]{getCountStatement()}},
\code{\link[=getSccsSettings]{getSccsSettings()}}
}
\concept{Extraction}
