% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfControlCaseSeries.R
\name{getSccsColDefs}
\alias{getSccsColDefs}
\title{getSccsColDefs}
\usage{
getSccsColDefs()
}
\value{
A column definition list
}
\description{
Create col defs for SCCS
}
\details{
Returns a named list of colDefs
}
