% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{getSettingsTable}
\alias{getSettingsTable}
\title{getSettingsTable}
\usage{
getSettingsTable(package, functionName, settings)
}
\arguments{
\item{package}{Name of installed R package of interest}

\item{functionName}{Name of exported function inside R package of interest}

\item{settings}{...}
}
\value{
Returns a tibble with the input details
}
\description{
This returns a tibble with the input details for the function of interest
in the package of interest plus add the user setting values for the inputs
adds a column called 'bold' that is TRUE if the user setting value is different
to the default
}
\details{
This returns a tibble with the input details
}
