% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{reportTableFormat}
\alias{reportTableFormat}
\title{reportTableFormat}
\usage{
reportTableFormat(
  table,
  groupBy = NULL,
  columns = NULL,
  elementId = NULL,
  caption,
  groupByButton = FALSE
)
}
\arguments{
\item{table}{data.frame or tibble with the data to present}

\item{groupBy}{column to group by (optional)}

\item{columns}{The column details (create default using defaultColumns())}

\item{elementId}{Element ID for the widget.}

\item{caption}{A table caption}

\item{groupByButton}{Whether to add a button that lets you group/ungroup rows in the table}
}
\value{
Details about all inputs into the functionName within R package of interest
}
\description{
create a grouped reactable::reactable
}
\details{
Returns a reactable::reactable
}
\seealso{
Other Helpers:
\code{\link[=formatCovariateSettings]{formatCovariateSettings()}},
\code{\link[=functionDefaults]{functionDefaults()}},
\code{\link[=getAllHelpDetails]{getAllHelpDetails()}},
\code{\link[=getAllHelpText]{getAllHelpText()}},
\code{\link[=getCohortDefinitionsFromJson]{getCohortDefinitionsFromJson()}},
\code{\link[=getConcepts]{getConcepts()}},
\code{\link[=getDemoLoc]{getDemoLoc()}},
\code{\link[=getFunctionFromArgName]{getFunctionFromArgName()}},
\code{\link[=getHelpText]{getHelpText()}},
\code{\link[=getNegativeControlsFromJson]{getNegativeControlsFromJson()}},
\code{\link[=getSettingsTable]{getSettingsTable()}},
\code{\link[=tagPrint]{tagPrint()}}
}
\concept{Helpers}
